Consolidate ITaskItemExtensions and ItemUtilities
