+++ src/Tasks/Microsoft.NET.Build.Tasks/Microsoft.NET.Build.Tasks.csproj
@@ -66,7 +66,7 @@
     <Compile Include="IContentAssetPreprocessor.cs" />
     <Compile Include="ProduceContentAssets.cs" />
     <Compile Include="GetNearestTargetFramework.cs" />
-    <Compile Include="ITaskItemExtensions.cs" />
+    <Compile Include="ItemUtilities.cs" />
     <Compile Include="CompilationOptionsConverter.cs" />
     <Compile Include="NuGetUtils.cs" />
     <Compile Include="GetAssemblyVersion.cs" />
